requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
pytz==2023.3
email-validator==2.1.0

